Issue 5787: Updated Basis Universal to v1.15 with rdo optimizations (#5789)
* Updated basisu to v1.15

* Added update script

* Added SSE support to encoder

* Renamed lib to graphics_transcoder_basisu

* compile fix

* Patch fixes

+++ engine/texc/src/texc_enc_default.cpp
@@ -94,7 +94,7 @@ namespace dmTexc
 
     static void DestroyDefault(Texture* texture)
     {
-        for(size_t i = 0; i < texture->m_Mips.Size(); ++i)
+        for(uint32_t i = 0; i < texture->m_Mips.Size(); ++i)
         {
             delete[] texture->m_Mips[i].m_Data;
         }
